fix(clippy): resolve all warnings in embeddenator-fs
- Replace deprecated criterion::black_box with std::hint::black_box
- Use clamp() instead of manual max().min() chains
- Rename StreamingIngester::new() to builder() for clarity
- Simplify wildcard patterns in match expressions
- Add XattrStorage type alias for complex HashMap type
- Fix field reassignment with struct literal initialization
- Add allow attribute for assertions_on_constants in tests
- Fix needless_range_loop with iter_mut().enumerate()
- Remove unnecessary u32 casts in fuse_shim
- Fix bool comparison (== false) to use negation
- Remove redundant closures in error mapping
- Remove unused imports (Write, Read)

All tests pass (84/84), clippy clean with -D warnings.

+++ src/fs/fuse_shim.rs
@@ -1262,8 +1262,8 @@ impl fuser::Filesystem for EngramFS {
 
         // Determine file type from mode
         let file_type = mode & libc::S_IFMT;
-        let major = ((rdev >> 8) & 0xff) as u32;
-        let minor = (rdev & 0xff) as u32;
+        let major = (rdev >> 8) & 0xff;
+        let minor = rdev & 0xff;
 
         let result = match file_type {
             libc::S_IFCHR => self.add_device(&file_path, true, major, minor, Vec::new()),
